removed unused exports, some renaming
