+++ apps/one-drive/src/inngest/functions/data-protection/refresh-item.test.ts
@@ -1,7 +1,7 @@
 import { expect, test, describe, vi, beforeEach } from 'vitest';
 import { createInngestFunctionMock, spyOnElba } from '@elba-security/test-utils';
 import { NonRetriableError } from 'inngest';
-import * as getItemConnector from '@/connectors/share-point/get-item';
+import * as getItemConnector from '@/connectors/share-point/item';
 import { organisationsTable } from '@/database/schema';
 import { encrypt } from '@/common/crypto';
 import { db } from '@/database/client';
